Updated index.md with Spam-related Enhancements
The current CoC has proven at times to be inadequate with regards to Spam. After reading through the GoBridge CoC, it became apparent that it was a more comprehensive option, and by implementing its language on Dealing with Spam, the TechMasters community can be better protected by decreasing the ambiguity of what constitutes Spam. Additionally added Acknowledgement for this section.